+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.6.0 (in development)
 
 - Added `Dsn` helper class to parse DSN strings
+- Locked database schema to MySql
 
 ## 2.5.8 (Sep 23, 2025)